motor_v1_validation/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
groq>=0.9.0
anthropic>=0.39.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
python-docx>=1.1.0
httpx>=0.27.0
